# docs/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A modular and scalable Go REST API template implementing Clean Architecture.",
        "title": "Go Clean Template API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/authz/permissions/{perm_id}": {
            "get": {
                "description": "Get permission details by unique ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "authz-permissions"
                ],
                "summary": "Get permission by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Permission ID",
                        "name": "perm_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/gct_internal_controller_restapi_response.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gct_internal_controller_restapi_response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gct_internal_controller_restapi_response.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Update permission details",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "authz-permissions"
                ],
                "summary": "Update permission",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Permission ID",
                        "name": "perm_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Permission update body",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.Permission"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/gct_internal_controller_restapi_response.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gct_internal_controller_restapi_response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gct_internal_controller_restapi_response.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete permission by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "authz-permissions"
                ],
                "summary": "Delete permission",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Permission ID",
                        "name": "perm_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/gct_internal_controller_restapi_response.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gct_internal_controller_restapi_response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gct_internal_controller_restapi_response.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/authz/roles/{role_id}/permissions/{perm_id}": {
            "post": {
                "description": "Add a permission to a role",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "authz-roles"
                ],
                "summary": "Add permission to role",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Role ID",
                        "name": "role_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Permission ID",
                        "name": "perm_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/gct_internal_controller_restapi_response.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gct_internal_controller_restapi_response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gct_internal_controller_restapi_response.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Remove a permission from a role",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "authz-roles"
                ],
                "summary": "Remove permission from role",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Role ID",
                        "name": "role_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Permission ID",
                        "name": "perm_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/gct_internal_controller_restapi_response.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gct_internal_controller_restapi_response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gct_internal_controller_restapi_response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/authz/scope": {
            "get": {
                "tags": [
                    "authz-scopes"
                ],
                "summary": "Get scope",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Path",
                        "name": "path",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Method",
                        "name": "method",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
...
        "/sessions": {
            "get": {
                "description": "List all active sessions for the current user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "sessions"
                ],
                "summary": "Get active sessions",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Offset",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Limit",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/gct_internal_controller_restapi_response.SuccessResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/gct_internal_controller_restapi_response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gct_internal_controller_restapi_response.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a session for user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "sessions"
                ],
                "summary": "Create a new session",
                "parameters": [
                    {
                        "description": "Session creation query",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.Session"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/gct_internal_controller_restapi_response.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/domain.Session"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gct_internal_controller_restapi_response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gct_internal_controller_restapi_response.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "domain.Permission": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "parent_id": {
                    "type": "string"
                }
            }
        },
...
        "domain.Session": {
            "type": "object",
            "properties": {
                "browser": {
                    "type": "string"
                },
                "browser_version": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "data": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "device_id": {
                    "type": "string"
                },
                "device_name": {
                    "type": "string"
                },
                "device_type": {
                    "$ref": "#/definitions/domain.SessionDeviceType"
                },
                "expires_at": {
                    "type": "string"
                },
                "fcm_token": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "ip_address": {
                    "type": "string"
                },
                "last_activity": {
                    "type": "string"
                },
                "os": {
                    "type": "string"
                },
                "os_version": {
                    "type": "string"
                },
                "refresh_token_hash": {
                    "type": "string"
                },
                "revoked": {
                    "type": "boolean"
                },
                "updated_at": {
                    "type": "string"
                },
                "user_agent": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "domain.SessionDeviceType": {
            "type": "string",
            "enum": [
                "DESKTOP",
                "MOBILE",
                "TABLET",
                "BOT",
                "TV"
            ],
            "x-enum-varnames": [
                "DeviceTypeDesktop",
                "DeviceTypeMobile",
                "DeviceTypeTablet",
                "DeviceTypeBot",
                "DeviceTypeTV"
            ]
        },
...
        "gct_internal_controller_restapi_response.ErrorDetail": {
            "type": "object",
            "properties": {
                "category": {
                    "type": "string",
                    "example": "DATA"
                },
                "code": {
                    "type": "string",
                    "example": "RESOURCE_NOT_FOUND"
                },
                "details": {
                    "type": "string",
                    "example": "The user with the ID '12345' does not exist in our records."
                },
                "message": {
                    "type": "string",
                    "example": "The requested resource was not found."
                },
                "method": {
                    "type": "string",
                    "example": "GET"
                },
                "path": {
                    "type": "string",
                    "example": "/api/v1/users/12345"
                },
                "retry_after": {
                    "description": "seconds",
                    "type": "integer",
                    "example": 5
                },
                "retryable": {
                    "type": "boolean",
                    "example": false
                },
                "severity": {
                    "description": "Enhanced fields",
                    "type": "string",
                    "example": "MEDIUM"
                },
                "suggestion": {
                    "type": "string",
                    "example": "Please check our documentation."
                },
                "timestamp": {
                    "type": "string",
                    "example": "2023-12-08T12:30:45Z"
                }
            }
        },
        "gct_internal_controller_restapi_response.ErrorResponse": {
            "type": "object",
            "properties": {
                "documentation_url": {
                    "type": "string",
                    "example": "https://developer.mozilla.org/en-US/docs/Web/HTTP/Status"
                },
                "error": {
                    "$ref": "#/definitions/gct_internal_controller_restapi_response.ErrorDetail"
                },
                "requestId": {
                    "type": "string",
                    "example": "a1b2c3d4-e5f6-7890-g1h2-i3j4k5l6m7n8"
                },
                "status": {
                    "type": "string",
                    "example": "error"
                },
                "statusCode": {
                    "type": "integer",
                    "example": 404
                }
            }
        },
        "gct_internal_controller_restapi_response.SuccessResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string",
                    "example": "Operation completed successfully"
                },
                "meta": {},
                "status": {
                    "description": "\"success\" or \"error\"",
                    "type": "string",
                    "example": "success"
                },
                "statusCode": {
                    "type": "integer",
                    "example": 200
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}